backend/requirements.txt
psycopg2-binary==2.9.6
gunicorn==21.2.0

amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
attrs==23.2.0
billiard==4.2.0
celery==5.4.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
cryptography==43.0.0
defusedxml==0.8.0rc2
distro==1.9.0
Django==5.0.7
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-cors-headers==4.4.0
django-filter==24.2
django-redis==5.4.0
django-templated-mail==1.1.1
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
drf-nested-routers==0.94.1
drf-spectacular==0.27.2
h11==0.14.0
html2image==2.0.4.3
httpcore==1.0.5
httpx==0.27.0
humanize==4.10.0
idna==3.7
inflection==0.5.1
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kombu==5.3.7
Levenshtein==0.25.1
Markdown==3.6
MarkupSafe==2.1.5
oauthlib==3.2.2
openai==1.37.1
pillow==10.4.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.8.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-Levenshtein==0.25.1
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
rapidfuzz==3.9.5
redis==5.0.7
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.19.1
six==1.16.0
sniffio==1.3.1
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
tornado==6.4.1
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
websocket-client==1.8.0